refactor: Move RecordStreamSourceStore out of inspections package
